Add test for versioned GC
